requirements.txt
# ARGUS — Python Dependencies

# Core
fastapi>=0.110,<1.0
uvicorn[standard]>=0.27,<1.0
pydantic>=2.5,<3.0
python-dotenv>=1.0,<2.0
httpx>=0.27,<1.0
structlog>=24.1,<25.0

# Kafka
confluent-kafka>=2.3,<3.0

# ML / Prediction
xgboost>=2.0,<3.0
shap>=0.44,<1.0
numpy>=1.26,<2.0
scikit-learn>=1.4,<2.0

# CV / Computer Vision (Rogesh's services)
# ultralytics>=8.1
# segment-anything
# opencv-python>=4.9

# PINN / Virtual Sensor Mesh + Causal GNN
torch>=2.2,<3.0

# TimescaleDB / PostgreSQL
asyncpg>=0.29,<1.0

# Signal processing
scipy>=1.12,<2.0

# Prediction fast-track
joblib>=1.3,<2.0
pandas>=2.1,<3.0
redis>=5.0,<6.0

# ACN Node (offline edge)
playsound>=1.3,<2.0
# twilio>=9.0,<10.0  # optional — uncomment for IVR calls

# ── Phase 2 ──────────────────────────────────────────────

# Causal Engine — GNN (optional, NumPy SEM fallback exists)
# torch-geometric>=2.5,<3.0  # optional — uncomment if CUDA available

# Evacuation RL
gymnasium>=0.29,<1.0
# stable-baselines3>=2.2,<3.0  # optional — heuristic fallback exists

# Federated Learning (uses numpy + torch)
# No extra deps — base64 weight transport over HTTP

# CHORUS NLP (keyword-based, no heavy NLP model needed)
# No extra deps

# MIRROR hydrological sim (uses numpy + scipy)
# No extra deps
